requirements.txt
alembic==1.18.3
annotated-types==0.7.0
astral==3.2
attrs==25.4.0
Authlib==1.6.6
blinker==1.9.0
brevo-python==1.2.0
cattrs==25.3.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.4
dotenv==0.9.9
Flask==3.1.2
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flatbuffers==25.9.23
geographiclib==2.1
geopy==2.4.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.188.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-genai
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httplib2==0.31.2
idna==3.11
itsdangerous==2.2.0
jh2==5.0.10
Jinja2==3.1.6
Mako==1.3.10
Markdown==3.10.1
MarkupSafe==3.0.3
niquests==3.17.0
numpy==2.4.1
openmeteo_requests==1.7.5
openmeteo_sdk==1.25.0
pandas==3.0.0
pillow==12.1.0
platformdirs==4.5.1
proto-plus==1.27.0
protobuf==5.29.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
qh3==1.5.6
requests==2.32.5
pycountry==24.6.1
requests-cache==1.2.1
retry-requests==2.0.0
six==1.17.0
SQLAlchemy==2.0.46
tqdm==4.67.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
url-normalize==2.2.1
urllib3==2.6.3
urllib3-future==2.15.901
wassima==2.0.4
Werkzeug==3.1.5
stripe
gunicorn
aerofiles
stripe
